refactor(workspace): split create.rs into DDD modules
Break apart the workspace create.rs file into several domain-driven
design files for better separation of concerns:

- editor.rs: EditorType enum and editor launching logic
- path.rs: Workspace path generation and sanitization
- data.rs: Issue data copying for workspace setup
- create.rs: Simplified orchestration layer

This follows the existing DDD patterns used in the item/ module.
